requirements.txt
django
djangorestframework
django-rest-auth
django-allauth
openpyxl
reportlab
simplejson
django-debug-toolbar


